### 2024-08-08 #7087 @editorialbot
[PRE REVIEW]: PYLIWC: Linguistic Inquiry and Word Count in Python
**Submitting author:** @camille1 (<a href="http://orcid.org/0000-0002-0815-9046">Camille LACAN</a>)
**Repository:** https://github.com/camille1/pyliwc
**Branch with paper.md** (empty if default branch):
**Version:** 0.0.1
