### 2024-08-12 #7087 @logological
@seyyaw @HLasse @alicia-ziying-yang @dpalmasan Would you be willing and able to review this manuscript for the Journal of Open Source Software?
