### 2024-08-17 #7087 @seyyaw
@logological I am new to this kind of review. Does the review also include `code review`? [This](https://raw.githubusercontent.com/openjournals/joss-papers/joss.07087/joss.07087/10.21105.joss.07087.pdf) is the paper to review right?

Let me know and I am happy to help!
