### 2024-08-19 #7087 @logological
@seyyaw Reviewing for JOSS is fairly easy -- we prepare a checklist for you that you can work through.  You'll need to examine both the paper and the code (which may involve installing and running the software in order to do some basic checks).  The entire review takes place via GitHub issues.  Any issues you find should be reported in the review issue, or raised as separate issues in the repository of the code/paper under review.  The authors then address those issues (by discussing them with you and/or making the necessary changes to the code/paper) so that you're able to proceed with the checklist.

Since you say you're happy to help, I'll go ahead and add you as a reviewer and we can begin the review.  This will take place in a separate GitHub issue of which you will be notified.  Please let me know if you have any questions!
