### 2024-08-20 #7119 @HLasse
@logological I have started my review, but am unsure how to continue/whether this is suitable for JOSS.

The software is mainly a wrapper around the CLI from a proprietary application ([LIWC](https://www.liwc.app/)) with some additional features. While I definitely see the use and merit of such a package, it cannot be used without an active license for LIWC. So, although this package is open source, it is deeply dependent on closed-source software which requires payment to access.

I am unable to perform a proper review of the functionality of the package as I do not have a license for LIWC. @logological If you think the package is still suited for JOSS you might need to find a different reviewer with a LIWC license. My apologies for not noticing before accepting!
