### 2024-08-20 #7119 @logological
@HLasse My apologies; my initial reading of the paper also did not lead me to understand that a LIWC licence was required to test the software.  My interpretation of the article summary is that the only thing proprietary about LIWC is the software, for which PYLIWC provides a free replacement.  The paper does not seem to state one way or another whether the data (dictionaries) used by the LIWC software are also under a proprietary licence.  But even if the LIWC dictionaries are proprietary, the article indicates that PYLIWC can be used with "custom dictionaries", which I assume might include freely available ones.

@camille1, could you please clarify these points, both here and in the paper itself?  Could you please also indicate how the reviewers of this paper can test the functionality of PYLIWC if they do not already hold a LIWC licence?  For example, is there an LIWC instance to which you can provide temporary, confidential access without violating the terms of the licence?  Or is there a free custom dictionary you can provide for testing purposes?
