### 2024-08-21 #7119 @logological
@camille1: While I'm at it, I've noticed a couple inconsistencies that I'd ask you to resolve:

1. The paper states that PYLIWC is distributed under the MIT licence, but the `LICENSE` file in the code repository contains the text of the Apache Software License.
2. The software's name is variously spelled "PYLIWC", "PyLIWC", and "pyliwc" in the paper and in various files in the code repository.
