+++ CHANGELOG.md
@@ -1,0 +1,9 @@
+# Release notes
+
+<!-- do not remove -->
+
+## 0.0.2
+
+### changes
+
+- update MIT License
