+++ settings.ini
@@ -1,16 +1,10 @@
 [DEFAULT]
-# All sections below are required unless otherwise specified.
-# See https://github.com/fastai/nbdev/blob/master/settings.ini for examples.
-
-### Python library ###
 repo = pyliwc
-lib_name = %(repo)s
+lib_name = pyliwc
 version = 0.0.1
 min_python = 3.7
-license = apache2
+license = MIT
 black_formatting = False
-
-### nbdev ###
 doc_path = _docs
 lib_path = pyliwc
 nbs_path = nbs
@@ -18,29 +12,26 @@ recursive = True
 tst_flags = notest
 put_version_in_init = True
 readme_nb = readme.ipynb
-
-### Docs ###
 branch = main
 custom_sidebar = True
-doc_host = https://%(user)s.github.io
-doc_baseurl = /%(repo)s
-git_url = https://github.com/%(user)s/%(repo)s
-title = %(lib_name)s
-
-### PyPI ###
+doc_host = https://camillelacan.github.io
+doc_baseurl = /pyliwc
+git_url = https://github.com/camillelacan/pyliwc
+title = pyliwc
 audience = Developers
 author = C.Lacan
 author_email = camille.lacan@univ-perp.fr
-copyright = 2024 onwards, %(author)s
+copyright = 2024 onwards, C.Lacan
 description = Linguistic Inquiry and Word Count (LIWC) in Python
-keywords =  LIWC "Linguistic Inquiry" "Textual Analysis" "Language Style Matching"
+keywords = LIWC "Linguistic Inquiry" "Textual Analysis" "Language Style Matching"
 language = English
 status = 3
 user = camillelacan
-
-### Optional ###
 requirements = pandas matplotlib
 dev_requirements = fastcore
-# console_scripts =
-# conda_user = 
-# package_data =
+allowed_metadata_keys = 
+allowed_cell_metadata_keys = 
+jupyter_hooks = True
+clean_ids = True
+custom_quarto_yml = False
+
