+++ CHANGELOG.md
@@ -1,9 +1,3 @@
 # Release notes
 
 <!-- do not remove -->
-
-## 0.0.2
-
-### changes
-
-- update MIT License
