+++ settings.ini
@@ -1,7 +1,7 @@
 [DEFAULT]
 repo = pyliwc
 lib_name = pyliwc
-version = 0.0.1
+version = 0.0.2
 min_python = 3.7
 license = MIT
 black_formatting = False
