+++ conda/pyliwc/meta.yaml
@@ -1,0 +1,149 @@
+package:
+  name: pyliwc
+  version: 0.0.2
+source:
+  sha256: e4160262662b398cbc096d7256046145f9d70f7da400e749bf66913eaeb64fb0
+  url: https://files.pythonhosted.org/packages/4f/41/d118c8d4c53a74cc4f69b5291317000cb2b9a15da82185beacbb4dcc4617/pyliwc-0.0.2.tar.gz
+about:
+  description: "\n\n![CI](https://github.com/camille1/pyliwc/actions/workflows/test.yaml/badge.svg)\n\
+    \n# Welcome to `pyliwc`\n\n> Linguistic Inquiry and Word Count* in Python\n\n\
+    # Overview\n\n`pyliwc` is a Python package designed to provide an interface for\n\
+    analyzing text using the LIWC (Linguistic Inquiry and Word Count) tool.\nThis\
+    \ package allows users to interact with the LIWC CLI from within\nPython, offering\
+    \ features for processing various data formats,\nperforming linguistic style matching,\
+    \ and analyzing narrative arcs in\ntext data. It can handle folders, text files\
+    \ or `Pandas` dataframes.\n\nAs the LIWC dictionary is proprietary software, this\
+    \ requires that you\nhave installed the latest version of the LIWC software on\
+    \ your machine,\nwith an activated licence (academic licence).\n\n# Manifest\n\
+    \nThe LIWC (Linguistic Inquiry and Word Count) software by James W.\nPennebaker,\
+    \ Roger J. Booth, and Martha E. Francis has been instrumental\nfor countless researchers\
+    \ in analyzing linguistic and psycholinguistic\ndata. Linguistic Inquiry and Word\
+    \ Count (LIWC) is the gold standard of\ndictionary-based approaches for analyzing\
+    \ word use. It can be used to\nstudy a single individual, groups of people over\
+    \ time, or all of social\nmedia.However, LIWC has traditionally been available\
+    \ through software,\nnecessitating the usage of an outside software to the Python\n\
+    environement.\n\nRecognizing the growing popularity of Python in the scientific\
+    \ and\nresearch community, there is an important need for researchers and data\n\
+    scientists for integrating LIWC directly into their Python workflows.\nThus, `pyliwc`\
+    \ brings (many of) the functionality of LIWC to a wider\naudience without the\
+    \ need use the LIWC application as GUI. `pyliwc` is\nopen-source, released under\
+    \ the MIT license, and is designed to enable\nresearchers to perform sophisticated\
+    \ linguistic analysis directly in\nPython.\n\n# Features\n\nThe package offers\
+    \ a wide range of features, including:\n\n1.  **LIWC Text Analysis:**\n\n- Analyze\
+    \ text data from various sources, including CSV files,\n  directories, Pandas\
+    \ DataFrames, and individual strings.\n- Supports internal dictionaries (e.g.,\
+    \ LIWC22, LIWC2015) as well as\n  adhoc dictionaries\n- Output results directly\
+    \ in a convenient Pandas DataFrame for easy\n  integration with other data processing\
+    \ tools.\n\n2.  **Linguistic Style Matching (LSM):**\n\n- Perform person and group-level\
[…]
