### 2024-08-29 #7119 @camille1
@logological I have clarified these points in the paper, specifying that the use of pyliwc requires the proprietary software LIWC in order to respect the copyright. pyliwc is open source (MIT license), but the software is not.  (see section licensing requirements in the paper).

I have fixed the two issues mentioned (MIT license, and correct spelling of “pyliwc” consistently in the documentation and the paper.)

I can share temporary and privately my LIWC license with reviewers @HLasse
