+++ quickstart.html.md
@@ -1,0 +1,437 @@
+# Quickstart
+
+
+<!-- WARNING: THIS FILE WAS AUTOGENERATED! DO NOT EDIT! -->
+
+Here’s a quickstart guide, explaining how to use `pyliwc` perform text
+analysis.
+
+[`Liwc`](https://camillelacan.github.io/pyliwc/core.html#liwc) is the
+main class for interacting with the LIWC CLI.
+
+``` python
+liwc = Liwc(liwc_cli_path='LIWC-22-cli', threads=None, verbose=True) 
+```
+
+**Parameters**: - liwc_cli_path (str): Path to the LIWC CLI executable.
+Default is ‘LIWC-22-cli’. On WSL, it is required to add .exe at the end
+‘LIWC-22-cli.exe’, - threads (int, optional): Number of threads to use.
+Defaults to the number of CPU cores minus one. - verbose (bool,
+optional): If True, display printing such as progress bar for large
+files. Defaults to False.
+
+The main Class of `Pyliwc` is
+[`Liwc`](https://camillelacan.github.io/pyliwc/core.html#liwc).
+
+<div class="alert alert-info">
+
+Note: Make sure that you have `LIWC-22` software running on your
+computer — it is required for using all methods
+
+</div>
+
+<!-- ![liwc22_interface.jpg](images/liwc22_interface.jpg) -->
+
+<img src="images/liwc22_interface.jpg" width="350">
+
+**Analyze a string using the default dictionary (LIWC22)**
+
+``` python
+from pyliwc import Liwc
+
+liwc = Liwc('LIWC-22-cli.exe')
+
+text = "This is a sample text for LIWC analysis."
+
+r = liwc.analyze_string_to_json(text)
+```
+
+    Picked up JAVA_TOOL_OPTIONS: -Dfile.encoding=UTF-8
+
+``` python
+{
+    'Segment': 1,
+    'WC': 8,
+    'Analytic': 89.52,
+    'Clout': 40.06,
+    'Authentic': 15.38,
+    'Tone': 20.23,
+    'WPS': 8,
+    'BigWords': 12.5,
+    'Dic': 100,
+    'Linguistic': 62.5,
+    'function': 50,
+    'pronoun': 12.5,
+    'ppron': 0,
+    'i': 0,
+    'we': 0,
+    'you': 0,
+    ... ,
+    'AllPunc': 12.5,
+    'Period': 12.5,
+    'Comma': 0,
+    'QMark': 0,
+    'Exclam': 0,
+    'Apostro': 0,
+    'OtherP': 0,
+    'Emoji': 0
+}
+
+```
+
+**Analyzing Texts from a `pandas` DataFrame**
+
+``` python
+import pandas as pd
+df = pd.read_csv('../data/US-president.csv')
+```
+
+``` python
+from pyliwc import Liwc
+
+liwc = Liwc('LIWC-22-cli.exe')
+
+liwc.analyze_df(df.Text, liwc_dict="LIWC22")
+```
+
+<div>
+<style scoped>
+    .dataframe tbody tr th:only-of-type {
+        vertical-align: middle;
+    }
+&#10;    .dataframe tbody tr th {
+        vertical-align: top;
+    }
+&#10;    .dataframe thead th {
+        text-align: right;
+    }
+</style>
+
+<table class="dataframe" data-quarto-postprocess="true" data-border="1">
+<thead>
+<tr class="header" style="text-align: right;">
[…]
