+++ how-to-use.html.md
@@ -1,0 +1,517 @@
+# How to use
+
+
+<!-- WARNING: THIS FILE WAS AUTOGENERATED! DO NOT EDIT! -->
+
+## Initialization
+
+[`Liwc`](https://camillelacan.github.io/pyliwc/core.html#liwc) is the
+main class for interacting with the LIWC CLI.
+
+``` python
+liwc = Liwc(liwc_cli_path='LIWC-22-cli', threads=None, verbose=True) 
+```
+
+**Parameters**: - liwc_cli_path (str): Path to the LIWC CLI executable.
+Default is ‘LIWC-22-cli’. On WSL, it is required to add .exe at the end
+‘LIWC-22-cli.exe’, - threads (int, optional): Number of threads to use.
+Defaults to the number of CPU cores minus one. - verbose (bool,
+optional): If True, display printing such as progress bar for large
+files. Defaults to False.
+
+The main Class of `Pyliwc` is
+[`Liwc`](https://camillelacan.github.io/pyliwc/core.html#liwc).
+
+<div class="alert alert-info">
+
+Note: Pyliwc requires that the LIWC-22 application is running on your
+computer. Please ensure that you currently have a copy of the LIWC-22
+user interface open.
+
+</div>
+
+<!-- ![liwc22_interface.jpg](images/liwc22_interface.jpg) -->
+
+<img src="images/liwc22_interface.jpg" width="350">
+
+## Methods
+
+### LIWC analyze
+
+`analyze_df`
+
+Analyze text data from a Pandas DataFrame using LIWC.
+
+``` python
+
+from pyliwc import Liwc
+
+liwc = Liwc('LIWC-22-cli.exe')
+
+
+result_df = liwc.analyze_df(text_df, return_input=False, liwc_dict='LIWC22')
+```
+
+**Example**
+
+``` python
+import pandas as pd
+data = {'Text': ["I am happy", "I feel sad today"]}
+df = pd.DataFrame(data)
+result_df = liwc.analyze_df(df.Text)
+```
+
+`analyze_csv`
+
+Analyze text data from a CSV file using LIWC.
+
+``` python
+liwc.analyze_csv(input_file, output_location, row_id_indices, column_indices, liwc_dict='LIWC22')
+```
+
+**Example**
+
+``` python
+liwc.analyze_csv("input.csv", "output.csv", "0", "1")
+```
+
+`analyze_folder`
+
+Analyze all text files in a specified folder using LIWC.
+
+``` python
+liwc.analyze_folder(input_folder, output_location, liwc_dict='LIWC22')
+```
+
+**Example**
+
+``` python
+liwc.analyze_folder('D:/Downloads/texts_folder', 'D:/Downloads')
+```
+
+`analyze_string`
+
+Analyze a single string using LIWC.
+
+``` python
+liwc.analyze_string(input_string, output_location, liwc_dict='LIWC22')
+```
+
+**Example**
+
+``` python
+liwc.analyze_string("I am feeling great", "D:/Downloads")
+```
+
+`analyze_string_to_json`
+
+Analyze a single string and return the result as a JSON object.
+
+``` python
+liwc.analyze_string_to_json(input_string, liwc_dict='LIWC22')
+```
+
+**Example**
+
+``` python
+result = liwc.analyze_string_to_json("This is an example string")
+print(result)
+
[…]
