+++ robots.txt
@@ -1,0 +1,1 @@
+Sitemap: https://camillelacan.github.io/pyliwc/sitemap.xml
