### 2024-08-29 #7119 @logological
@HLasse @seyyaw Would the author's offer to use his LIWC licence for the purpose of reviewing this submission work for you?
