### 2024-08-29 #7119 @seyyaw
@logological Sharing the license sounds like a good idea, but I'm not sure about the ethical implications of sharing the LIWC license. Do they permit sharing with third parties @camille1 ? I'll look into whether there's a free license available for academic use, but I won't be able to do this until after mid-September because I'm traveling next week.
