### 2024-08-29 #7119 @camille1
@logological @HLasse @seyyaw There is no free LIWC license, and the version I have is a paid academic version.

as explained, I can share my license, the disadvantage is that it's a license for one user at a time (so we have to agree on the day each reviewer needs it).
