### 2024-09-20 #7119 @samhforbes
Hi @camille1 thanks for your patience. Looking at the fact that LIWC is not open source, as well as the relatively small size of this package, we were unable to agree that PYLIWC is in scope. I want to stress that this does not mean the software is not useful, simply that it does not meet our requirements for JOSS.

Thanks @HLasse, @seyyaw and @logological for reviewing and editing
